major restructure; split rules into seperate files
