+++ .gitignore
@@ -4,3 +4,5 @@ dist/
 sample_cases/
 wasm-toolchain
 coverage
+.vscode/settings.json
+reference
